% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_multi_polysomes.R
\name{graph_multi_polysomes}
\alias{graph_multi_polysomes}
\title{Graph multiple polysome profiles on one graph}
\usage{
graph_multi_polysomes(x, min_position)
}
\arguments{
\item{x}{list of data frames created by read_polysomes function.}

\item{min_position}{minimum position from top of gradient at which to start graphing from}
}
\description{
This function takes a list of data frames generated by using read_polysomes with lapply and
graphs each sample as a different color on a single graph.
}
