% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_single_polysomes.R
\name{graph_single_polysomes}
\alias{graph_single_polysomes}
\title{Graph a single polysome profile}
\usage{
graph_single_polysomes(x, min_position)
}
\arguments{
\item{x}{data frame created by read_polysomes function.}

\item{min_position}{minimum position from top of gradient at which to start graphing from}
}
\description{
This function graphs data from a single sample with fractions shown in alternating colors. It takes a data frame created by read_polysomes
and the minimum distance from the top of the gradient at which you want to start graphing from.
}
