% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metadata}
\alias{read_metadata}
\title{Get metadata from polysome file}
\usage{
read_metadata(x)
}
\arguments{
\item{x}{file path to a polysome csv file.}
}
\description{
This function retrieves the metadata from the polysome csv. Warning: the metadata for these files is only
partially controlled by the machine, the user enters some of it, hence it may not be correct. Please keep
careful notes or ensure that you entered it correctly on the machine. Additionally, in the event of manual
polysome collection, the fraction volume and number of samples will be intentionally wrong.
}
