% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_polysomes.R
\name{read_polysomes}
\alias{read_polysomes}
\title{Read a polysome profiling csv from the BioComp machine}
\usage{
read_polysomes(filepath)
}
\arguments{
\item{filepath}{Path to the input file.}
}
\description{
This function loads a csv generated from the biocomp machine, discarding any metadata in the process.
There is a separate function read_metadata for getting the metadata.
It then cleans the data to make it more amenable to the computer. It does not remove or alter data.
It will add a filename column to the data, which is derived from the last element in the path.
This function takes a single path to a csv. Use with lapply to make a list of files.
}
\details{
The machine only puts the fraction number in the row where the signal is sent to the fractionator, meaning
that every row with a fraction number separates two fractions. For example, the first fraction is all rows
above the row with a 1 in it, and then fraction 2 is between 2 and 3, and so on. The last fraction is called "A"
for "air". This function copies this column to fraction_points and fills the fraction column with values,
converting the A to a number in the process.

This function throws a warning if one of the values in the
fraction_number column was duplicated, as may be the case with the A fraction, perhaps because you let
go of the button and pressed it again when you were getting the last bit out of the machine.
}
