% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_facet_polysomes.R
\name{graph_facet_polysomes}
\alias{graph_facet_polysomes}
\title{Graph multiple polysome profiles on separate graphs, but in one plot.}
\usage{
graph_facet_polysomes(x, min_position)
}
\arguments{
\item{x}{list of data frames created by using read_polysomes with lapply.}

\item{min_position}{minimum position from top of gradient at which to start graphing from}
}
\description{
This uses facet_wrap to graph many samples on the same plot but in different panels. It takes a list of read_polysome data frames
and a minimum position. It labels fraction positions. It has fixed scales.
}
